Chaged tiffRead to deal with non-scanimage tiffs
If a metadata output argument is requested but no metadata is found in
the tiff file, tiffRead will now output a fake scanimage structure and
issue a warning. This should allow the Acq2p object to be used with
non-scanimage tiffs, but it is a dirty workaround because the fake
metadata does not correspond to the actual tiff file.