tiffReader permutes output
scanimage tiffreader fills dimensions based on how they are acquired,
not displayed. This means that the 'x' axis of the image is the first
('y') axis of the matrix. This is confusing, and counter to what
ordinary readers do, which is to match image and matrix dimensions.
tiffReader now handles this for us